requirements.txt
django
django-tastypie
django-autofixture
